refactor(ui): share one menu row between the utility menus
The room rail and the lobby sidebar each carried a private copy, and they
had already drifted — only the sidebar's could tint a destructive action.
A third menu is about to reach the same destinations from the sign-in
header, so the row moves to core/ui with the superset behaviour.
